Reorg Bayesian optimizers to a single class per file (#342)
* reorg bayesian optimizers to a single class per file

* adding a test to make sure all implementations of optimizers are listed in the optimizer type enum

* do the same for space adapters

* tweaks to use abstract classes for mypy in a few other places

* check that those lists are non-empty

* extend the mypy type-abstract ignores
