+++ mlos_bench/mlos_bench/launcher.py
@@ -12,9 +12,9 @@ command line.
 
 import logging
 import argparse
-from typing import Any, Dict, Iterable, List, Optional, Tuple
+from typing import Any, Dict, Iterable, List, Optional, Tuple, Type
 
-from mlos_bench.util import BaseTypes
+from mlos_bench.util import BaseTypeVar
 from mlos_bench.tunables.tunable_groups import TunableGroups
 from mlos_bench.environments.base_environment import Environment
 
@@ -218,8 +218,7 @@ class Launcher:
         if args_optimizer is None:
             return OneShotOptimizer(
                 self.tunables, self._parent_service, self.global_config)
-        optimizer = self._load(Optimizer, args_optimizer)
-        assert isinstance(optimizer, Optimizer)
+        optimizer = self._load(Optimizer, args_optimizer)   # type: ignore[type-abstract]
         return optimizer
 
     def _load_storage(self, args_storage: Optional[str]) -> Storage:
@@ -233,20 +232,25 @@ class Launcher:
             from mlos_bench.storage.sql.storage import SqlStorage
             return SqlStorage(self.tunables, self._parent_service,
                               {"drivername": "sqlite", "database": ":memory:"})
-        storage = self._load(Storage, args_storage)
-        assert isinstance(storage, Storage)
+        storage = self._load(Storage, args_storage)     # type: ignore[type-abstract]
         return storage
 
-    def _load(self, cls: type, json_file_name: str) -> BaseTypes:
+    def _load(self, cls: Type[BaseTypeVar], json_file_name: str) -> BaseTypeVar:
         """
         Create a new instance of class `cls` from JSON configuration.
+
+        Note: For abstract types, mypy will complain at the call site.
+        Use "# type: ignore[type-abstract]" to suppress the warning.
+        See Also: https://github.com/python/mypy/issues/4717
         """
         class_config = self._config_loader.load_config(json_file_name)
         assert isinstance(class_config, Dict)
-        return self._config_loader.build_generic(
+        ret = self._config_loader.build_generic(
             base_cls=cls,
             tunables=self.tunables,
             service=self._parent_service,
             config=class_config,
             global_config=self.global_config
         )
+        assert isinstance(ret, cls)
+        return ret
